+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 - Added updateModel
 - Added getModel
 - Added new possible value 'inactive' for status in Model
+- Added CreateAppClientOptions type for createAppClient. You may now create app clients for both the code and
+client_credentials OAuth2 grant.